move icon cell
