+++ frontend/pages/SoftwarePage/components/IconCell/IconCell.tsx
