+++ frontend/pages/SoftwarePage/components/IconCell/_styles.scss
