+++ frontend/pages/SoftwarePage/components/IconCell/index.ts
